pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "pm"
description = "List and manage available projects"
dynamic = ["version"]
readme = "README.md"
authors = [
    {name = "Stanislav Sabev", email = "bezraboten.34@gmail.com"},
]
license = {file =  "LICENSE"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 1 - Planning",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
]
keywords = ["cli", "project"]
dependencies = [
    "gitpython>=3.1.42",
]

[tool.setuptools.dynamic]
version = {attr = "pm.__version__"}

[tool.setuptools.package-dir]
pm = "pm"

[project.optional-dependencies] 
dev = [
    "coverage>=7.4.4",
    "mock>=5.1.0",
    "mypy>=1.9.0",
    "pytest>=8.1.1",
    "ruff>=0.3.4",
    "types-mock>=5.1.0.20240311",
    "typing-extensions>=4.10.0",
]

[project.scripts]
    pm = "pm.cli:app"

[project.urls]
Homepage = "https://github.com/stanislavsabev/pm"
Repository = "https://github.com/stanislavsabev/pm.git"
"Bug Tracker" = "https://github.com/stanislavsabev/pm/issues"
Documentation = "https://github.com/stanislavsabev/pm/blob/main/README.md"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.mypy]
files = "pm/*.py"
strict = true
check_untyped_defs = true
ignore_missing_imports = true
scripts_are_modules = true
exclude = '''(?x)(
    tests/*
)'''

[tool.ruff]
line-length = 99
indent-width = 4

[tool.ruff.lint]
select = ["D", "E4", "E7", "E9", "F", "B", "SIM", "I001"]
ignore = ["D104", "D107", "D100"]

fixable = ["ALL"]
unfixable = ["B"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,testing,build,docs,tools}/*" = ["D", "E402"]

[tool.bandit]
targets = "."
exclude = "tests"

